fix(dsp): correct demo_sig dtype/seq_len bugs and crop axis transpose
- demo_sig: gauss/uniform now return float32 (matches docstring + tests).
- _demo_sig_pac: cast seq_len to int so np.zeros doesn't reject float t_sec*fs.
- crop: when axis != -1, the post-windowing array has one extra leading dim
  (n_windows). Old inverse-permutation used the original ndim and crashed with
  'axes don't match array'. Replace with an explicit remap that places n_windows
  at the original axis position and keeps window_length trailing.
- test_crop_different_axis: fixture is internally inconsistent (signal_3d[:50]
  is still (4,5,100), so axis=0 yields 1 window of length 10, not the 5 the
  assert expects). Mark xfail with a clear reason rather than silently masking.

+++ src/scitex_dsp/_demo_sig.py
@@ -122,13 +122,19 @@ def demo_sig(
 
     if sig_type == "uniform":
         return (
-            np.random.uniform(low=-0.5, high=0.5, size=(batch_size, n_chs, len(tt))),
+            np.random.uniform(
+                low=-0.5, high=0.5, size=(batch_size, n_chs, len(tt))
+            ).astype(np.float32),
             tt,
             fs,
         )
 
     elif sig_type == "gauss":
-        return np.random.randn(batch_size, n_chs, len(tt)), tt, fs
+        return (
+            np.random.randn(batch_size, n_chs, len(tt)).astype(np.float32),
+            tt,
+            fs,
+        )
 
     elif sig_type == "meg":
         return (
@@ -221,7 +227,7 @@ def _demo_sig_pac(
     -------
         np.array: Generated signals with shape (batch_size, n_chs, n_segments, seq_len).
     """
-    seq_len = t_sec * fs
+    seq_len = int(t_sec * fs)
     t = np.arange(seq_len) / fs
     if verbose:
         print(f"Generating signal with length: {seq_len}")
